Move assert_paths_exist into fs_helpers
